tf2: Fixes compiler warnings on OS X
This exploited a gcc specific extension and is not
C++ standard compliant. There used to be a "fix"
for OS X which no longer applies. I think it is ok
to use this as an int instead of a double, but
another way to fix it would be to use a define.

+++ tf2/include/tf2/buffer_core.h
@@ -87,7 +87,7 @@ class BufferCore
 {
 public:
   /************* Constants ***********************/
-  static const double DEFAULT_CACHE_TIME = 10.0;  //!< The default amount of time to cache data in seconds
+  static const int DEFAULT_CACHE_TIME = 10;  //!< The default amount of time to cache data in seconds
   static const uint32_t MAX_GRAPH_DEPTH = 1000UL;  //!< The default amount of time to cache data in seconds
 
   /** Constructor
